Difference between abscissa and ordinate of points A & B
