+++ hardware/include/KNNsw_reg.v
@@ -1,4 +1,5 @@
 //START_TABLE sw_reg
-`SWREG_W(KNN_RESET,          1, 0) //Timer soft reset
-`SWREG_W(KNN_ENABLE,         1, 0) //Timer enable
-
+`SWREG_W(KNN_RESET,          1, 0) // KNN soft reset
+`SWREG_W(KNN_ENABLE,         1, 0) // KNN enable
+`SWREG_W(KNN_X,        32, 0) // KNN point A
+`SWREG_W(KNN_Y,         32, 0) // KNN point B
