+++ hardware/include/iob_knn.vh
@@ -1,5 +1,5 @@
-`define KNN_ADDR_W 3  //address width
-`define KNN_WDATA_W 1 //write data width
+`define KNN_ADDR_W 8  //address width
+`define KNN_WDATA_W 32 //write data width
 `ifndef DATA_W
  `define DATA_W 32      //cpu data width
 `endif
