+++ software/embedded/embedded.mk
@@ -3,4 +3,4 @@
 include $(KNN_DIR)/software/software.mk
 
 #embeded sources
-#SRC+=$(KNN_SW_DIR)/embedded/
+SRC+=$(KNN_SW_DIR)/embedded/iob_knn.c
