+++ software/iob_knn.h
@@ -5,4 +5,6 @@ void knn_reset();
 void knn_start();	
 void knn_stop();
 void knn_init( int base_address);	
-
+void knn_set_x(int x);
+void knn_set_y(int y);
+signed long int knn_get_dist();
